requirements.txt
# 운세 Shorts 자동 생성기 - 의존성
# Python 3.14 호환 (pre-built wheel 사용, 소스 빌드 방지)
streamlit>=1.40.0
numpy>=2.0.0
moviepy==1.0.3
pillow>=10.2.0
opencv-python>=4.9.0.80
openai>=1.12.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0
python-dotenv>=1.0.1
pandas>=2.2.0
